pyproject.toml
[tool.poetry]
name = "home-work-drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
pillow = "^11.0.0"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
black = "^24.10.0"
mypy = "^1.14.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
ipython = "^8.31.0"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
